+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.2 (unreleased)
 ----------------
 
+- Added new marked pair counters marked_npairs_3d and marked_npairs_xy_z
+
 - Added new pair counters npairs_s_mu, npairs_jackknife_3d, npairs_projected, npairs_xy_z
 
 - Replaced npairs function with npairs_3d, which has the exact same API but is 30%-50x faster.